Convert to setuptools, update readme to reflect new install process
