+++ hardware_in_the_loop/software/README.md
@@ -2,7 +2,7 @@
 
 ## Intro
 
-This directory stores all code used to control and test with Olin Electric Motorsports' hardware-in-the-loop harness. It is organized into modules, tests, scripts, and artifacts (see below).
+This directory stores all code used to control and test with Olin Electric Motorsports' hardware-in-the-loop harness. It is organized into modules, tests, scripts, docs, and artifacts (see below).
 
 Note: All paths in this README file are given relative to the `software` directory.
 
@@ -20,6 +20,10 @@ All tests go here! That includes both unit tests (that test this code itself), a
 
 This is where functional scripts go, including our setup script.
 
+### Docs
+
+This is where our documentation lives! I've tried to document in as many as places as possible, but the docs folder contains the documentation generated by [sphinx](https://www.sphinx-doc.org/en/master/).
+
 ### Artifacts
 
 This is where all other files live. That includes our `config.ini` file, our can spec, all pin information, and logs, among other things.
@@ -30,7 +34,7 @@ This is a work in progress! It is not at all functional yet, but the current goa
 
 ## Setup
 
-DISCLAIMER: This script requires root permissions  to edit some files, and will also append to your `.bashrc` file.
+DISCLAIMER: This process requires root permissions to edit some files.
 
 ### Prerequisites
 * A linux system. Due to the hardware interfaces that this project relies on, I STRONGLY recommend against using a virtual machine; this project will not work out of the box on a VM.
@@ -38,27 +42,41 @@ DISCLAIMER: This script requires root permissions  to edit some files, and will 
 
 ### Installation
 
-0. (OPTIONAL) This setup script will automatically create symbolic links to certain hardware it finds. For example, if you have an arduino plugged in, this script creates a rule that makes your arduino accessible at `/dev/arduino` all the time, so you don't have to guess whether it is at `/dev/ACM0` or `/dev/ACM1` or any other location anymore. Paths to hardware in this repository are **hard-coded** and **assume these symbolic links exist**, so these rules are necessary for consistent performance. If you have an arduino, I recommend plugging it in now.
-
 1. If you haven't yet, clone this repository onto your computer
 
     `$ git clone https://github.com/olin-electric-motorsports/AdvancedResearch.git`
 
-2. Navigate to the folder with our setup script (the script must be run from the `scripts` directory because it includes relative paths!)
+2. To perform software installation and setup, navigate to the software directory
     
-    `$ cd AdvancedResearch/hardware_in_the_loop/software/scripts`
+    `$ cd AdvancedResearch/hardware_in_the_loop/software`
 
-3. Run the setup script
+3. (Optional) If you want to setup a virtual envrionment for this project (I reccomend it), now is the time. If you don't know how to do this, read [here](https://realpython.com/intro-to-pyenv/)
 
-    `$ sudo python setup.py` or `$ sudo python3 setup.py`
+4. Use pip to install the project and its dependencies on your computer
 
-4. Restart your computer.
+    `$ pip install -e .`
+
+5. Now that the software is installed, its time to configure hardware. For now, the only hardware we can connect to is an arduino. Go ahead and plug that in now. If you don't have an arduino, thats fine; you just won't be able to run all unit tests.
+
+6. Now that your arduino is plugged in, verify that your computer recognizes it with 
+
+    `$ lsusb`
+    
+    You should see a your arduino listed as an option. If you don't do not proceed to step 6; try unplugging it and plugging it in again.
+    
+7. Navigate to our scripts directory and run our hardware setup script
+
+    `$ cd scripts && sudo python hardware_setup.py`
+
+    This script creates a udev rule which generates a permanent symbolic link to your arduino at `/dev/arduino`, so we never need to guess where to find it. The software relies on these symbolic links, so we will need them to work to interact with hardware.
+
+8. Restart your computer.
 
     `$ sudo shutdown -r now`
 
-    The reason you have to reboot is because this script adds udev rules, which create symbolic links to our hardware (see above). It also edits your `~/.bashrc` file. If you know what you're doing and really don't want to reboot, you can instead try 
+    It is generally recommended to restart to reload udev rules. If you're in a rush, you can try 
     
-    `$ source ~/.bashrc && sudo udevadm control --reload-rules`,
+    `$ sudo udevadm control --reload-rules`,
 
     but I generally recommend restarting just to be sure.
 
@@ -161,6 +179,7 @@ You'll see that `artifacts/pin_info.csv` has 9 columns. The `IOController` pytho
 
 * [socketcan docs](https://python-can.readthedocs.io/en/master/interfaces/socketcan.html)
 
+* [pyenv](https://realpython.com/intro-to-pyenv/)
 
 
 If you go to Olin, the fastest way to get help is probably to hit up our Slack channel, `#hardware_in_the_loop`. Otherwise, you can email awenstrup@olin.edu
